bugfix #2461 - figure handles for naming are now cast to double do be HG2 compatible
git-svn-id: svn+ssh://svn.fcdonders.nl/home/svnroot/fieldtrip/trunk@9778 0cf7c7f0-3615-4144-b4e6-68da3bce3cd0

+++ ft_layoutplot.m
@@ -129,9 +129,9 @@ else
 end
 
 if ~isempty(dataname)
-  set(gcf, 'Name', sprintf('%d: %s: %s', gcf, funcname, dataname));
+  set(gcf, 'Name', sprintf('%d: %s: %s', double(gcf), funcname, dataname));
 else
-  set(gcf, 'Name', sprintf('%d: %s', gcf, funcname));
+  set(gcf, 'Name', sprintf('%d: %s', double(gcf), funcname));
 end
 set(gcf, 'NumberTitle', 'off');
