+++ ft_multiplotTFR.m
@@ -628,7 +628,7 @@ if isempty(get(gcf, 'Name'))
   end
   
   if isempty(cfg.figurename)
-    set(gcf, 'Name', sprintf('%d: %s: %s', gcf, funcname, dataname));
+    set(gcf, 'Name', sprintf('%d: %s: %s', double(gcf), funcname, dataname));
     set(gcf, 'NumberTitle', 'off');
   else
     set(gcf, 'name', cfg.figurename);
